+++ src/rate.js
@@ -11,7 +11,7 @@
 function rate ( obj, req, res, next ) {
 	var headers = ["x-ratelimit-limit", "x-ratelimit-remaining", "x-ratelimit-reset"],
 	    results = obj.rate( req ),
-		valid   = results.shift(),
+	    valid   = results.shift(),
 	    config  = obj.server.config.rate;
 
 	array.each( headers, function ( i, idx ) {