+++ CHANGELOG.md
@@ -1,10 +1,9 @@
 # Change Log
 
 ## 0.4.0
-- Added `Facebook` authentication via `passport`
+- Added `Facebook`, & `Twitter` authentication via `passport`
 - Fixed the `RegExp` used to retrieve the rate limit id from the `authorization` header
-- Moved authentication gateway middleware into `zuul()`
-- Fixed what `zuul()` tests
+- Moved authentication gateway middleware into `zuul()`, & fixed what it tests
 
 ## 0.3.6
 - Upgraded turtle.io to 2.3.0